# swagger.yml
openapi: 3.0.2
servers:
  - url: 'http://127.0.0.1:8080'
    description: local server
info:
  version: 1.0.0
  title: API client to cache specification
tags:
  - name: keys
    description: Operations with keys in cache
paths:
  /set:
    post:
      tags:
        - keys
      summary: Set new key and value
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  description: string key value
                  type: string
                value:
                  description: value as string, array or object
                  oneOf:
                    - type: string
                    - type: array
                      items: {}
                    - type: object
                ttl:
                  description: key ttl
                  type: string
              required:
                - key
                - value
                - ttl
            examples:
              string:
                summary: string value example
                value:
                  key: name
                  value: Ivan
                  ttl: 5m
              array:
                summary:  array value example
                value:
                  key: nicknames
                  value: [Ivan1999, Ivashka, MadSkorpion]
                  ttl: 5m
              map:
                summary: map value example
                value:
                  key: pets
                  value:
                    dog:
                      Polkan
                    cat:
                      Murka
                  ttl: 5m
      responses:
        '200':
          description: Successful set operation
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResp'
        '500':
          description: Internal error in cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResp'
  /get/{key}:
    get:
      tags:
        - keys
      summary: Get key value
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            example: name
      responses:
        '200':
          description: Key value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValueResp'
        '500':
          description: Internal error in cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResp'
  /getListElem/{key}/{index}:
    get:
      tags:
        - keys
      summary: For array value - get element by array index
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            example: nicknames
        - name: index
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
      responses:
        '200':
          description: element value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValueResp'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResp'
        '500':
          description: Internal error in cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResp'
  /getMapElemValue/{key}/{mapKey}:
    get:
      tags:
        - keys
      summary: For map value - get element key value
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            example: pets
        - name: mapKey
          in: path
          required: true
          schema:
            type: string
            example: cat
      responses:
        '200':
          description: Element value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValueResp'
        '500':
          description: Internal error in cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResp'
  /remove/{key}:
    delete:
      tags:
        - keys
      summary: Remove element from cache by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            example: name
      responses:
        '200':
          description: Successful remove
        '500':
          description: Internal error in cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResp'
  /keys:
    get:
      tags:
        - keys
      summary: Get all keys from cache
      responses:
        '200':
          description: Keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeysResp'
        '500':
          description: Internal error in cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResp'
components:
  schemas:
    ErrorResp:
      type: object
      properties:
        error:
          type: string
    ValueResp:
      type: object
      properties:
        value:
          oneOf:
            - type: string
            - type: array
              items: {}
            - type: object
    KeysResp:
      type: object
      properties:
        keys:
          type: array
          items:
            type: string